Update index.html
Added manual input of nSequence function. By default, all input nSequences for the transaction are set to the same value 0xfffffffd (2 ^ 32-3=4294967293), which activates the RBF (Replace By Fee) function. When the given transaction fee is relatively low and the transaction cannot be packaged for a long time, a higher fee transaction can be initiated to replace the previously broadcasted low fee transaction; The absolute time lock (nLockTime) has been activated also, which can be used to construct transaction strings for inheritance.